+++ src/main/java/blackjack/controller/BlackjackController.java
@@ -2,7 +2,7 @@ package blackjack.controller;
 
 import blackjack.domain.card.Cards;
 import blackjack.domain.card.Deck;
-import blackjack.domain.card.Result;
+import blackjack.domain.Result;
 import blackjack.domain.player.Dealer;
 import blackjack.domain.player.Player;
 import blackjack.view.InputView;
